+++ sdk/scvmm/azure-resourcemanager-scvmm/src/main/java/com/azure/resourcemanager/scvmm/fluent/VirtualMachineInstancesClient.java
@@ -577,14 +577,13 @@ public interface VirtualMachineInstancesClient {
      * The operation to power off (stop) a virtual machine instance.
      * 
      * @param resourceUri The fully qualified Azure Resource manager identifier of the resource.
-     * @param body The content of the action request.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      * @return the {@link SyncPoller} for polling of long-running operation.
      */
     @ServiceMethod(returns = ReturnType.LONG_RUNNING_OPERATION)
-    SyncPoller<PollResult<Void>, Void> beginStop(String resourceUri, StopVirtualMachineOptions body);
+    SyncPoller<PollResult<Void>, Void> beginStop(String resourceUri);
 
     /**
      * Implements the operation to stop a virtual machine.
@@ -608,13 +607,12 @@ public interface VirtualMachineInstancesClient {
      * The operation to power off (stop) a virtual machine instance.
      * 
      * @param resourceUri The fully qualified Azure Resource manager identifier of the resource.
-     * @param body The content of the action request.
      * @throws IllegalArgumentException thrown if parameters fail the validation.
      * @throws com.azure.core.management.exception.ManagementException thrown if the request is rejected by server.
      * @throws RuntimeException all other wrapped checked exceptions if the request fails to be sent.
      */
     @ServiceMethod(returns = ReturnType.SINGLE)
-    void stop(String resourceUri, StopVirtualMachineOptions body);
+    void stop(String resourceUri);
 
     /**
      * Implements the operation to stop a virtual machine.