+++ sdk/scvmm/azure-resourcemanager-scvmm/src/main/java/com/azure/resourcemanager/scvmm/models/CloudInventoryItem.java
@@ -20,26 +20,6 @@ public final class CloudInventoryItem extends InventoryItemProperties {
      */
     private InventoryType inventoryType = InventoryType.CLOUD;
 
-    /*
-     * Gets the tracked resource id corresponding to the inventory resource.
-     */
-    private String managedResourceId;
-
-    /*
-     * Gets the UUID (which is assigned by Vmm) for the inventory item.
-     */
-    private String uuid;
-
-    /*
-     * Gets the Managed Object name in Vmm for the inventory item.
-     */
-    private String inventoryItemName;
-
-    /*
-     * Provisioning state of the resource.
-     */
-    private ProvisioningState provisioningState;
-
     /**
      * Creates an instance of CloudInventoryItem class.
      */
@@ -57,53 +37,12 @@ public final class CloudInventoryItem extends InventoryItemProperties {
     }
 
     /**
-     * Get the managedResourceId property: Gets the tracked resource id corresponding to the inventory resource.
-     * 
-     * @return the managedResourceId value.
-     */
-    @Override
-    public String managedResourceId() {
-        return this.managedResourceId;
-    }
-
-    /**
-     * Get the uuid property: Gets the UUID (which is assigned by Vmm) for the inventory item.
-     * 
-     * @return the uuid value.
-     */
-    @Override
-    public String uuid() {
-        return this.uuid;
-    }
-
-    /**
-     * Get the inventoryItemName property: Gets the Managed Object name in Vmm for the inventory item.
-     * 
-     * @return the inventoryItemName value.
-     */
-    @Override
-    public String inventoryItemName() {
-        return this.inventoryItemName;
-    }
-
-    /**
-     * Get the provisioningState property: Provisioning state of the resource.
-     * 
-     * @return the provisioningState value.
-     */
-    @Override
-    public ProvisioningState provisioningState() {
-        return this.provisioningState;
-    }
-
-    /**
      * Validates the instance.
      * 
      * @throws IllegalArgumentException thrown if the instance is not valid.
      */
     @Override
     public void validate() {
-        super.validate();
     }
 
     /**
@@ -132,13 +71,14 @@ public final class CloudInventoryItem extends InventoryItemProperties {
                 reader.nextToken();
 
                 if ("managedResourceId".equals(fieldName)) {
-                    deserializedCloudInventoryItem.managedResourceId = reader.getString();
+                    deserializedCloudInventoryItem.withManagedResourceId(reader.getString());
                 } else if ("uuid".equals(fieldName)) {
-                    deserializedCloudInventoryItem.uuid = reader.getString();
+                    deserializedCloudInventoryItem.withUuid(reader.getString());
                 } else if ("inventoryItemName".equals(fieldName)) {
-                    deserializedCloudInventoryItem.inventoryItemName = reader.getString();
+                    deserializedCloudInventoryItem.withInventoryItemName(reader.getString());
                 } else if ("provisioningState".equals(fieldName)) {
-                    deserializedCloudInventoryItem.provisioningState = ProvisioningState.fromString(reader.getString());
+                    deserializedCloudInventoryItem
+                        .withProvisioningState(ProvisioningState.fromString(reader.getString()));
                 } else if ("inventoryType".equals(fieldName)) {
                     deserializedCloudInventoryItem.inventoryType = InventoryType.fromString(reader.getString());
                 } else {