+++ sdk/scvmm/azure-resourcemanager-scvmm/src/main/java/com/azure/resourcemanager/scvmm/models/VirtualMachineTemplateInventoryItem.java
@@ -40,26 +40,6 @@ public final class VirtualMachineTemplateInventoryItem extends InventoryItemProp
      */
     private String osName;
 
-    /*
-     * Gets the tracked resource id corresponding to the inventory resource.
-     */
-    private String managedResourceId;
-
-    /*
-     * Gets the UUID (which is assigned by Vmm) for the inventory item.
-     */
-    private String uuid;
-
-    /*
-     * Gets the Managed Object name in Vmm for the inventory item.
-     */
-    private String inventoryItemName;
-
-    /*
-     * Provisioning state of the resource.
-     */
-    private ProvisioningState provisioningState;
-
     /**
      * Creates an instance of VirtualMachineTemplateInventoryItem class.
      */
@@ -113,53 +93,12 @@ public final class VirtualMachineTemplateInventoryItem extends InventoryItemProp
     }
 
     /**
-     * Get the managedResourceId property: Gets the tracked resource id corresponding to the inventory resource.
-     * 
-     * @return the managedResourceId value.
-     */
-    @Override
-    public String managedResourceId() {
-        return this.managedResourceId;
-    }
-
-    /**
-     * Get the uuid property: Gets the UUID (which is assigned by Vmm) for the inventory item.
-     * 
-     * @return the uuid value.
-     */
-    @Override
-    public String uuid() {
-        return this.uuid;
-    }
-
-    /**
-     * Get the inventoryItemName property: Gets the Managed Object name in Vmm for the inventory item.
-     * 
-     * @return the inventoryItemName value.
-     */
-    @Override
-    public String inventoryItemName() {
-        return this.inventoryItemName;
-    }
-
-    /**
-     * Get the provisioningState property: Provisioning state of the resource.
-     * 
-     * @return the provisioningState value.
-     */
-    @Override
-    public ProvisioningState provisioningState() {
-        return this.provisioningState;
-    }
-
-    /**
      * Validates the instance.
      * 
      * @throws IllegalArgumentException thrown if the instance is not valid.
      */
     @Override
     public void validate() {
-        super.validate();
     }
 
     /**
@@ -189,14 +128,14 @@ public final class VirtualMachineTemplateInventoryItem extends InventoryItemProp
                 reader.nextToken();
 
                 if ("managedResourceId".equals(fieldName)) {
-                    deserializedVirtualMachineTemplateInventoryItem.managedResourceId = reader.getString();
+                    deserializedVirtualMachineTemplateInventoryItem.withManagedResourceId(reader.getString());
                 } else if ("uuid".equals(fieldName)) {
-                    deserializedVirtualMachineTemplateInventoryItem.uuid = reader.getString();
+                    deserializedVirtualMachineTemplateInventoryItem.withUuid(reader.getString());
                 } else if ("inventoryItemName".equals(fieldName)) {
-                    deserializedVirtualMachineTemplateInventoryItem.inventoryItemName = reader.getString();
+                    deserializedVirtualMachineTemplateInventoryItem.withInventoryItemName(reader.getString());
                 } else if ("provisioningState".equals(fieldName)) {
-                    deserializedVirtualMachineTemplateInventoryItem.provisioningState
-                        = ProvisioningState.fromString(reader.getString());
+                    deserializedVirtualMachineTemplateInventoryItem
+                        .withProvisioningState(ProvisioningState.fromString(reader.getString()));
                 } else if ("inventoryType".equals(fieldName)) {
                     deserializedVirtualMachineTemplateInventoryItem.inventoryType
                         = InventoryType.fromString(reader.getString());